# make array comparison faster