get rid of redundant comparisons

in array comparison, if match has been found - no need to proceed
further - break at that point and proceed with next element

```
go test -bench .
BenchmarkBigArrays-8    	    1000	   1811571 ns/op
BenchmarkBigArrays2-8   	    1000	   1822322 ns/op
PASS
ok  	github.com/edast/jsonpatch	4.010s
```

after adding break statements:
```
go test -bench .
BenchmarkBigArrays-8    	    2000	    911185 ns/op
BenchmarkBigArrays2-8   	    2000	    923130 ns/op
PASS
ok  	github.com/edast/jsonpatch	3.877s
```